requirements.txt
Pillow==2.5.1
selenium==2.42.1
scipy==0.14.0
numpy==1.8.1
